Add CloudFront, S3, and Lambda@Edge Terraform module (#10)
Adds a Terraform module that encapsulates the resources necessary to support an S3 bucket fronted by a CloudFront distribution. Communication between CloudFront and S3 is protected by an origin access identity policy.

In addition, a Lambda@Edge function is associated with the CloudFront distribution. It users a JavaScript function to add the `X-Frame-Options` header to responses viewer responses.